refactor di of files and separate each concept in object class for more readability
