--- app/src/main/java/com/paris_2/aflami/di/HomeRemoteDataSourceModule.kt
@@ -1,24 +1,0 @@
-package com.paris_2.aflami.di
-
-import com.repository.media.GenresDataSourceImpl
-import com.repository.media.MediaDataSourceImpl
-import com.repository.media.datasource.remote.GenresRemoteDataSource
-import com.repository.media.datasource.remote.MediaRemoteDataSource
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object HomeRemoteDataSourceModule {
-    @Provides
-    @Singleton
-    fun provideGenresRemoteDataSource(impl: GenresDataSourceImpl): GenresRemoteDataSource = impl
-
-    @Provides
-    @Singleton
-    fun provideMediaRemoteDataSource(impl: MediaDataSourceImpl): MediaRemoteDataSource = impl
-}
-
